+++ rtl/core/neorv32_cpu.vhd
@@ -5,7 +5,9 @@
 -- # * neorv32_cpu.vhd                   - CPU top entity                                          #
 -- #   * neorv32_cpu_alu.vhd             - Arithmetic/logic unit                                   #
 -- #   * neorv32_cpu_bus.vhd             - Instruction and data bus interface unit                 #
--- #   * neorv32_cpu_cp_muldiv.vhd       - MULDIV co-processor                                     #
+-- #   * neorv32_cpu_cp_bitmanip.vhd     - Bit-manipulation co-processor ('B')                     #
+-- #   * neorv32_cpu_cp_fpu.vhd          - Single-precision FPU co-processor ('F')                 #
+-- #   * neorv32_cpu_cp_muldiv.vhd       - Integer multiplier/divider co-processor ('M')           #
 -- #   * neorv32_cpu_ctrl.vhd            - CPU control and CSR system                              #
 -- #     * neorv32_cpu_decompressor.vhd  - Compressed instructions decoder                         #
 -- #   * neorv32_cpu_regfile.vhd         - Data register file                                      #
@@ -62,10 +64,11 @@ entity neorv32_cpu is
     CPU_EXTENSION_RISCV_B        : boolean := false; -- implement bit manipulation extensions?
     CPU_EXTENSION_RISCV_C        : boolean := false; -- implement compressed extension?
     CPU_EXTENSION_RISCV_E        : boolean := false; -- implement embedded RF extension?
+    CPU_EXTENSION_RISCV_F        : boolean := false; -- implement 32-bit floating-point extension?
     CPU_EXTENSION_RISCV_M        : boolean := false; -- implement muld/div extension?
     CPU_EXTENSION_RISCV_U        : boolean := false; -- implement user mode extension?
     CPU_EXTENSION_RISCV_Zicsr    : boolean := true;  -- implement CSR system?
-    CPU_EXTENSION_RISCV_Zifencei : boolean := true;  -- implement instruction stream sync.?
+    CPU_EXTENSION_RISCV_Zifencei : boolean := false; -- implement instruction stream sync.?
     -- Extension Options --
     FAST_MUL_EN                  : boolean := false; -- use DSPs for M extension's multiplier
     FAST_SHIFT_EN                : boolean := false; -- use barrel shifter for shift operations
@@ -121,27 +124,31 @@ end neorv32_cpu;
 architecture neorv32_cpu_rtl of neorv32_cpu is
 
   -- local signals --
-  signal ctrl       : std_ulogic_vector(ctrl_width_c-1 downto 0); -- main control bus
-  signal comparator : std_ulogic_vector(1 downto 0); -- comparator result
-  signal imm        : std_ulogic_vector(data_width_c-1 downto 0); -- immediate
-  signal instr      : std_ulogic_vector(data_width_c-1 downto 0); -- new instruction
-  signal rs1, rs2   : std_ulogic_vector(data_width_c-1 downto 0); -- source registers
-  signal alu_res    : std_ulogic_vector(data_width_c-1 downto 0); -- alu result
-  signal alu_add    : std_ulogic_vector(data_width_c-1 downto 0); -- alu address result
-  signal rdata      : std_ulogic_vector(data_width_c-1 downto 0); -- memory read data
-  signal alu_wait   : std_ulogic; -- alu is busy due to iterative unit
-  signal bus_i_wait : std_ulogic; -- wait for current bus instruction fetch
-  signal bus_d_wait : std_ulogic; -- wait for current bus data access
-  signal csr_rdata  : std_ulogic_vector(data_width_c-1 downto 0); -- csr read data
-  signal mar        : std_ulogic_vector(data_width_c-1 downto 0); -- current memory address register
-  signal ma_instr   : std_ulogic; -- misaligned instruction address
-  signal ma_load    : std_ulogic; -- misaligned load data address
-  signal ma_store   : std_ulogic; -- misaligned store data address
-  signal be_instr   : std_ulogic; -- bus error on instruction access
-  signal be_load    : std_ulogic; -- bus error on load data access
-  signal be_store   : std_ulogic; -- bus error on store data access
-  signal fetch_pc   : std_ulogic_vector(data_width_c-1 downto 0); -- pc for instruction fetch
-  signal curr_pc    : std_ulogic_vector(data_width_c-1 downto 0); -- current pc (for current executed instruction)
+  signal ctrl          : std_ulogic_vector(ctrl_width_c-1 downto 0); -- main control bus
+  signal comparator    : std_ulogic_vector(1 downto 0); -- comparator result
+  signal imm           : std_ulogic_vector(data_width_c-1 downto 0); -- immediate
+  signal instr         : std_ulogic_vector(data_width_c-1 downto 0); -- new instruction
+  signal rs1, rs2      : std_ulogic_vector(data_width_c-1 downto 0); -- source registers
+  signal alu_res       : std_ulogic_vector(data_width_c-1 downto 0); -- alu result
+  signal alu_add       : std_ulogic_vector(data_width_c-1 downto 0); -- alu address result
+  signal mem_rdata     : std_ulogic_vector(data_width_c-1 downto 0); -- memory read data
+  signal mem_wdata     : std_ulogic_vector(data_width_c-1 downto 0); -- memory write-data
+  signal alu_wait      : std_ulogic; -- alu is busy due to iterative unit
+  signal bus_i_wait    : std_ulogic; -- wait for current bus instruction fetch
+  signal bus_d_wait    : std_ulogic; -- wait for current bus data access
+  signal csr_rdata     : std_ulogic_vector(data_width_c-1 downto 0); -- csr read data
+  signal mar           : std_ulogic_vector(data_width_c-1 downto 0); -- current memory address register
+  signal ma_instr      : std_ulogic; -- misaligned instruction address
+  signal ma_load       : std_ulogic; -- misaligned load data address
+  signal ma_store      : std_ulogic; -- misaligned store data address
+  signal be_instr      : std_ulogic; -- bus error on instruction access
+  signal be_load       : std_ulogic; -- bus error on load data access
+  signal be_store      : std_ulogic; -- bus error on store data access
+  signal fetch_pc      : std_ulogic_vector(data_width_c-1 downto 0); -- pc for instruction fetch
+  signal curr_pc       : std_ulogic_vector(data_width_c-1 downto 0); -- current pc (for current executed instruction)
+  signal fpu_mem_wdata : std_ulogic_vector(data_width_c-1 downto 0); -- memory write-data form FPU
+  signal fpu_rm        : std_ulogic_vector(2 downto 0); -- FPU rounding mode
+  signal fpu_flags     : std_ulogic_vector(4 downto 0); -- FPU exception flags
 
   -- co-processor interface --
   signal cp_start  : std_ulogic_vector(7 downto 0); -- trigger co-processor i
@@ -174,8 +181,11 @@ begin
   -- A extension - only lr.w and sc.w are supported yet --
   assert not (CPU_EXTENSION_RISCV_A = true) report "NEORV32 CPU CONFIG WARNING! Atomic operations extension (A) only supports <lr.w> and <sc.w> instructions." severity warning;
 
-  -- Bit manipulation notifier --
-  assert not (CPU_EXTENSION_RISCV_B = true) report "NEORV32 CPU CONFIG WARNING! Bit manipulation extension (B) is still highly experimental (not ratified yet)." severity warning;
+  -- FIXME: Bit manipulation warning --
+  assert not (CPU_EXTENSION_RISCV_B = true) report "NEORV32 CPU CONFIG WARNING! Bit manipulation extension (B) is still HIGHLY EXPERIMENTAL (and spec. is not ratified yet)." severity warning;
+
+  -- FIXME: Floating-point extension warning --
+  assert not (CPU_EXTENSION_RISCV_F = true) report "NEORV32 CPU CONFIG WARNING! 32-bit floating-point extension (F) is WORK-IN-PROGRESS and NOT OPERATIONAL yet." severity warning;
 
   -- PMP regions check --
   assert not (PMP_NUM_REGIONS > 64) report "NEORV32 CPU CONFIG ERROR! Number of PMP regions <PMP_NUM_REGIONS> out of valid range (0..64)." severity error;
@@ -205,6 +215,7 @@ begin
     CPU_EXTENSION_RISCV_B        => CPU_EXTENSION_RISCV_B,        -- implement bit manipulation extensions?
     CPU_EXTENSION_RISCV_C        => CPU_EXTENSION_RISCV_C,        -- implement compressed extension?
     CPU_EXTENSION_RISCV_E        => CPU_EXTENSION_RISCV_E,        -- implement embedded RF extension?
+    CPU_EXTENSION_RISCV_F        => CPU_EXTENSION_RISCV_F,        -- implement 32-bit floating-point extension?
     CPU_EXTENSION_RISCV_M        => CPU_EXTENSION_RISCV_M,        -- implement muld/div extension?
     CPU_EXTENSION_RISCV_U        => CPU_EXTENSION_RISCV_U,        -- implement user mode extension?
     CPU_EXTENSION_RISCV_Zicsr    => CPU_EXTENSION_RISCV_Zicsr,    -- implement CSR system?
@@ -234,6 +245,9 @@ begin
     fetch_pc_o    => fetch_pc,    -- PC for instruction fetch
     curr_pc_o     => curr_pc,     -- current PC (corresponding to current instruction)
     csr_rdata_o   => csr_rdata,   -- CSR read data
+    -- FPU interface --
+    fpu_rm_o      => fpu_rm,      -- rounding mode
+    fpu_flags_i   => fpu_flags,   -- exception flags
     -- interrupts (risc-v compliant) --
     msw_irq_i     => msw_irq_i,   -- machine software interrupt
     mext_irq_i    => mext_irq_i,  -- machine external interrupt
@@ -271,7 +285,7 @@ begin
     clk_i  => clk_i,              -- global clock, rising edge
     ctrl_i => ctrl,               -- main control bus
     -- data input --
-    mem_i  => rdata,              -- memory read data
+    mem_i  => mem_rdata,          -- memory read data
     alu_i  => alu_res,            -- ALU result
     -- data output --
     rs1_o  => rs1,                -- operand 1
@@ -396,11 +410,40 @@ begin
   cp_valid(3)  <= cp_start(3); -- always assigned even if Zicsr extension is disabled to make sure CPU does not get stalled if there is an accidental access
 
 
-  -- Co-Processor 4..7: Not Implemented Yet -------------------------------------------------
+  -- Co-Processor 4: Single-Precision Floating-Point Unit ('F' Extension) -------------------
   -- -------------------------------------------------------------------------------------------
-  cp_result(4) <= (others => '0');
-  cp_valid(4)  <= '0';
-  --
+  neorv32_cpu_cp_fpu_inst_true:
+  if (CPU_EXTENSION_RISCV_F = true) generate
+    neorv32_cpu_cp_fpu_inst: neorv32_cpu_cp_fpu
+    port map (
+      -- global control --
+      clk_i     => clk_i,         -- global clock, rising edge
+      rstn_i    => rstn_i,        -- global reset, low-active, async
+      ctrl_i    => ctrl,          -- main control bus
+      start_i   => cp_start(4),   -- trigger operation
+      -- data input --
+      frm_i     => fpu_rm,        -- rounding mode
+      reg_i     => rs1,           -- rf source
+      mem_i     => mem_rdata,     -- memory read-data
+      -- result and status --
+      fflags_o  => fpu_flags,     -- exception flags
+      mem_o     => fpu_mem_wdata, -- memory write-data
+      res_o     => cp_result(4),  -- operation result
+      valid_o   => cp_valid(4)    -- data output valid
+    );
+  end generate;
+
+  neorv32_cpu_cp_fpu_inst_false:
+  if (CPU_EXTENSION_RISCV_F = false) generate
+    fpu_flags     <= (others => '0');
+    fpu_mem_wdata <= (others => '0');
+    cp_result(4)  <= (others => '0');
+    cp_valid(4)   <= cp_start(4); -- to make sure CPU does not get stalled if there is an accidental access
+  end generate;
+
+
+  -- Co-Processor 5..7: Not Implemented Yet -------------------------------------------------
+  -- -------------------------------------------------------------------------------------------
   cp_result(5) <= (others => '0');
   cp_valid(5)  <= '0';
   --
@@ -436,8 +479,8 @@ begin
     be_instr_o     => be_instr,       -- bus error on instruction access
     -- cpu data access interface --
     addr_i         => alu_add,        -- ALU.add result -> access address
-    wdata_i        => rs2,            -- write data
-    rdata_o        => rdata,          -- read data
+    wdata_i        => mem_wdata,      -- write data
+    rdata_o        => mem_rdata,      -- read data
     mar_o          => mar,            -- current memory address register
     d_wait_o       => bus_d_wait,     -- wait for access to complete
     --
@@ -473,6 +516,9 @@ begin
     d_bus_fence_o  => d_bus_fence_o,  -- fence operation
     d_bus_lock_o   => d_bus_lock_o    -- locked/exclusive access
   );
+
+  -- memory write data --
+  mem_wdata <= fpu_mem_wdata when ((CPU_EXTENSION_RISCV_F = true) and (ctrl(ctrl_bus_wd_sel_c) = '1')) else rs2;
 
   -- current privilege level --
   i_bus_priv_o <= ctrl(ctrl_priv_lvl_msb_c downto ctrl_priv_lvl_lsb_c);